0.6.0 — Protocol 18: Liquidity Pool operations
Build: Qt 6.11.1 / MSVC 2022 — 439/439 tests green, 63 test classes.

New
- CAP-38 XDR primitives: AssetType::POOL_SHARE, PoolID, LiquidityPoolType,
  LiquidityPoolConstantProductParameters, LiquidityPoolParameters,
  ChangeTrustAsset and TrustLineAsset unions.
- LiquidityPool helper: getLiquidityPoolID(type, assetA, assetB, fee)
  returning SHA-256 of the canonical parameters XDR, plus compareAssets()
  and the LIQUIDITY_POOL_FEE_V18 constant.
- C++ wrappers: ChangeTrustAsset / TrustLineAsset abstract bases with a
  plain-asset Wrapper subclass and a LiquidityPoolShare* subclass each.
- LiquidityPoolDepositOperation and LiquidityPoolWithdrawOperation
  (create() / build() / setSourceAccount() / getters). Operation::fromXdr
  factory extended for the two new opcodes.
- ChangeTrustOperation now accepts a ChangeTrustAsset (pool-share capable)
  in addition to the legacy Asset* ctor. New getLine() returns the wrapper
  for all 4 variants; getAsset() returns nullptr for the pool-share case.
- ChangeTrustOperationResponse gains the liquidity_pool_id field.

Changed
- stellar::ChangeTrustOp::line: Asset → ChangeTrustAsset. Wire format
  unchanged for the 3 pre-existing variants; only the POOL_SHARE variant
  uses the new union case.
- Operation union: ChangeTrustOp is now placement-new-managed (it
  contains a non-trivially-destructible ChangeTrustAsset). Internal
  Operation::fillChangeTrustOp() helper added.
- OperationResult union extended with the two new result codes so
  submit-response parsing tolerates LP-bearing transaction results.
- Build cleanup: cleared QVariant::Type / QVariant::type() deprecations
  and a few size_t→int/uint narrowing warnings on the Qt 6.11.1 / MSVC
  toolchain.

Deprecated
- None.

Compat: existing ChangeTrustOperation(Asset*, QString) callers compile
unchanged and produce byte-identical XDR. No public API removed. Version
bumped to 0.6.0.

+++ src/assettypecreditalphanum.cpp
@@ -27,7 +27,7 @@ const KeyPair &AssetTypeCreditAlphaNum::getIssuer() const{
 }
 
 uint AssetTypeCreditAlphaNum::hashCode() const{
-    return qHash(m_issuer->getAccountId() + this->getCode());
+    return static_cast<uint>(qHash(m_issuer->getAccountId() + this->getCode()));
 }
 
 bool AssetTypeCreditAlphaNum::equals(Asset *object) {
